backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
transformers==4.36.2
torch==2.1.2
python-dotenv==1.0.0
pillow==10.2.0